pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["TonyMM55 <tonymmoran@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
langchain = "^0.2.11"
langchain-openai = "^0.1.17"
langgraph = "^0.1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
